text-translator/requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=0.1.0
